fixed gitignore :)

+++ scripts/NET_FCN.py
@@ -67,7 +67,7 @@ class Net(nn.Module):# FCN Net class for semantic segmentation init generate net
                 RGBMean = [123.68,116.779,103.939]
                 RGBStd = [65,65,65]
                 if EvalMode==False: InpImages = torch.autograd.Variable(torch.from_numpy(Images.astype(float)), requires_grad=False).transpose(2,3).transpose(1, 2).type(torch.FloatTensor)
-                else: InpImages = torch.autograd.Variable(torch.from_numpy(Images.astype(float)),requires_grad=False).transpose(2, 3).transpose(1, 2).type(torch.HalfTensor)
+                else: InpImages = torch.autograd.Variable(torch.from_numpy(Images.astype(float)),requires_grad=False).transpose(2, 3).transpose(1, 2).type(torch.FloatTensor)
 
                 if self.UseGPU == True:
                     InpImages=InpImages.cuda()
